Implement signing for message digests
The functions `encryptMessageDigest` have been added to `DigitalSignature.php` in order to implement message digest signing functionality. The signed files will be saved in the newly created directory `client_signed_files`, which has been added to `.gitignore` along with an update in `KeyGenerator.php`. This was done in order to add an extra layer of security, providing authenticity, non-repudiation, and integrity checks for transmitted messages. `README.md` has been updated with information about the functions used in this process. Note that the `openssl_sign` function was not used, details in README.md.