Remove server signing comment, add message to Decryptor
In `DigitalSignature.php`, the hardcoded message, "Digitally signed by server", appended to the file content was removed as it was misleading. A message should not directly be stated as digitally signed.

Conversely, in `Decryptor.php`, a new comment has been added to `decryptedContent`, after the decryption process. This is to simulate a server replying to decrypted messages, adding a bit of realism to testing scenarios.